dd /exam/recommended-questions endpoint for recommendations

+++ src/main/java/com/pdd/pddai/dto/RecommendationQuestionDto.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 @Data
 public class RecommendationQuestionDto {
-    private int questionId;
+    private long questionId;
     private String questionText;
     private List<String> answers;
     private int correctAnswerIndex;
